Realização das funcões de envio das mensagens
