+++ whats-lab/src/compontents/Secaomensagem/SecaoMensagem.css
@@ -1,0 +1,23 @@
+.container-mensagem {
+  width: 500px;
+  height: 500px;
+  border: 1px solid black;
+}
+.input-usuario {
+  width: 90px;
+  height: 30px;
+  border: 1px solid red;
+}
+
+.input-mensagem {
+  width: 300px;
+  height: 30px;
+  border: 1px solid black;
+}
+
+button {
+  width: 100px;
+  height: 30px;
+
+  
+}
