--- whats-lab/src/compontents/secaoMensagem/mensagem/mensagem.js
@@ -1,2 +1,0 @@
-import React from 'react';
-
